requirements.txt
colorama
cycler
fonttools
joblib
kiwisolver
matplotlib
numpy
packaging
pandas
Pillow
pyparsing
python-dateutil
pytz
scikit-learn
scipy
six
threadpoolctl
torch
torchsummary
torchvision
tqdm
typing_extensions
